resolve sidebar
